gestion de deffaillance (ajout,modif,supprimer)

+++ src/main/java/com/group/transtubackend/repositories/DefaillanceRepository.java
@@ -2,12 +2,6 @@ package com.group.transtubackend.repositories;
 
 import com.group.transtubackend.entities.Defaillance;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 
-import java.time.LocalDate;
-import java.util.List;
-
-public interface DefaillanceRepository extends JpaRepository<Defaillance, Long> {
-
+public interface DefaillanceRepository extends JpaRepository<Defaillance, Integer> {
 }
